refactor view models into proper places and files
