+++ vividly-app/Controllers/ManageController.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin.Security;
 using vividly_app.Models;
+using vividly_app.ViewModels;
 
 namespace vividly_app.Controllers
 {
