+++ vividly-app/ViewModels/ConfigureTwoFactorViewModel.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace vividly_app.ViewModels
+{
+    public class ConfigureTwoFactorViewModel
+    {
+        public string SelectedProvider { get; set; }
+        public ICollection<System.Web.Mvc.SelectListItem> Providers { get; set; }
+    }
+}
