+++ vividly-app/ViewModels/ExternalLoginListViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace vividly_app.ViewModels
+{
+    public class ExternalLoginListViewModel
+    {
+        public string ReturnUrl { get; set; }
+    }
+}
