+++ vividly-app/ViewModels/FactorViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace vividly_app.ViewModels
+{
+    public class FactorViewModel
+    {
+        public string Purpose { get; set; }
+    }
+}
