+++ vividly-app/ViewModels/IndexViewModel.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+using Microsoft.AspNet.Identity;
+
+namespace vividly_app.ViewModels
+{
+    public class IndexViewModel
+    {
+        public bool HasPassword { get; set; }
+        public IList<UserLoginInfo> Logins { get; set; }
+        public string PhoneNumber { get; set; }
+        public bool TwoFactor { get; set; }
+        public bool BrowserRemembered { get; set; }
+    }
+}
