+++ vividly-app/ViewModels/ManageLoginsViewModel.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+using Microsoft.AspNet.Identity;
+using Microsoft.Owin.Security;
+
+namespace vividly_app.ViewModels
+{
+    public class ManageLoginsViewModel
+    {
+        public IList<UserLoginInfo> CurrentLogins { get; set; }
+        public IList<AuthenticationDescription> OtherLogins { get; set; }
+    }
+}
