+++ vividly-app/Views/Account/ExternalLoginConfirmation.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.ExternalLoginConfirmationViewModel
+﻿@model vividly_app.ViewModels.ExternalLoginConfirmationViewModel
 @{
     ViewBag.Title = "Register";
 }
