+++ vividly-app/Views/Account/ForgotPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.ForgotPasswordViewModel
+﻿@model vividly_app.ViewModels.ForgotPasswordViewModel
 @{
     ViewBag.Title = "Forgot your password?";
 }
