+++ vividly-app/Views/Account/Login.cshtml
@@ -1,5 +1,6 @@
 ﻿@using vividly_app.Models
-@model vividly_app.Models.LoginViewModel
+@using vividly_app.ViewModels
+@model vividly_app.ViewModels.LoginViewModel
 @{
     ViewBag.Title = "Log in";
 }
