+++ vividly-app/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.RegisterViewModel
+﻿@model vividly_app.ViewModels.RegisterViewModel
 @{
     ViewBag.Title = "Register";
 }
