+++ vividly-app/Views/Account/ResetPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.ResetPasswordViewModel
+﻿@model vividly_app.ViewModels.ResetPasswordViewModel
 @{
     ViewBag.Title = "Reset password";
 }
