+++ vividly-app/Views/Account/SendCode.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.SendCodeViewModel
+﻿@model vividly_app.ViewModels.SendCodeViewModel
 @{
     ViewBag.Title = "Send";
 }
