+++ vividly-app/Views/Account/VerifyCode.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.VerifyCodeViewModel
+﻿@model vividly_app.ViewModels.VerifyCodeViewModel
 @{
     ViewBag.Title = "Verify";
 }
