+++ vividly-app/Views/Account/_ExternalLoginsListPartial.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.ExternalLoginListViewModel
+﻿@model vividly_app.ViewModels.ExternalLoginListViewModel
 @using Microsoft.Owin.Security
 
 <h4>Use another service to log in.</h4>
