+++ vividly-app/Views/Manage/AddPhoneNumber.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.AddPhoneNumberViewModel
+﻿@model vividly_app.ViewModels.AddPhoneNumberViewModel
 @{
     ViewBag.Title = "Phone Number";
 }
