+++ vividly-app/Views/Manage/ChangePassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.ChangePasswordViewModel
+﻿@model vividly_app.ViewModels.ChangePasswordViewModel
 @{
     ViewBag.Title = "Change Password";
 }
