+++ vividly-app/Views/Manage/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.IndexViewModel
+﻿@model vividly_app.ViewModels.IndexViewModel
 @{
     ViewBag.Title = "Manage";
 }
