+++ vividly-app/Views/Manage/ManageLogins.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.ManageLoginsViewModel
+﻿@model vividly_app.ViewModels.ManageLoginsViewModel
 @using Microsoft.Owin.Security
 @{
     ViewBag.Title = "Manage your external logins";
