+++ vividly-app/Views/Manage/SetPassword.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.SetPasswordViewModel
+﻿@model vividly_app.ViewModels.SetPasswordViewModel
 @{
     ViewBag.Title = "Create Password";
 }
