+++ vividly-app/Views/Manage/VerifyPhoneNumber.cshtml
@@ -1,4 +1,4 @@
-﻿@model vividly_app.Models.VerifyPhoneNumberViewModel
+﻿@model vividly_app.ViewModels.VerifyPhoneNumberViewModel
 @{
     ViewBag.Title = "Verify Phone Number";
 }
